+++ lib/src/internal/completer_action.dart
@@ -12,4 +12,7 @@ class CompleterAction extends TimeAction {
   void execute() {
     complete.call();
   }
+
+  @override
+  void dispose() {}
 }