+++ lib/src/internal/repeat_event_action.dart
@@ -45,4 +45,7 @@ class RepeatEventAction extends TimeAction {
     }
     SimDartHelper.scheduleNextAction(sim: sim);
   }
+
+  @override
+  void dispose() {}
 }