+++ test/repeat_process_test.dart
@@ -1,26 +1,35 @@
 import 'package:simdart/simdart.dart';
 import 'package:test/test.dart';
 
-import 'track_tester.dart';
+import 'test_helper.dart';
 
 void main() {
+  late SimDart sim;
+  TestHelper helper = TestHelper();
+
+  setUp(() {
+    sim = SimDart(observer: helper);
+  });
+
   group('Repeat process', () {
     test('Simple', () async {
-      SimDart sim = SimDart(includeTracks: true);
-
       sim.repeatProcess(
           event: (context) async {},
           name: (start) => 'A$start',
           interval: Interval.fixed(fixedInterval: 1, untilTime: 2));
 
-      SimResult result = await sim.run();
-      TrackTester tt = TrackTester(result);
-      tt.test(['[0][A0][called]', '[1][A1][called]', '[2][A2][called]']);
+      await sim.run();
+      helper.test([
+        '[0][A0][called]',
+        '[0][A0][finished]',
+        '[1][A1][called]',
+        '[1][A1][finished]',
+        '[2][A2][called]',
+        '[2][A2][finished]'
+      ]);
     });
 
     test('Wait', () async {
-      SimDart sim = SimDart(includeTracks: true);
-
       sim.repeatProcess(
           event: (context) async {
             await context.wait(1);
@@ -28,25 +37,24 @@ void main() {
           name: (start) => 'A$start',
           interval: Interval.fixed(fixedInterval: 1, untilTime: 2));
 
-      SimResult result = await sim.run();
-
-      TrackTester tt = TrackTester(result);
-      tt.test([
+      await sim.run();
+      helper.test([
         '[0][A0][called]',
         '[0][A0][yielded]',
         '[1][A0][resumed]',
+        '[1][A0][finished]',
         '[1][A1][called]',
         '[1][A1][yielded]',
         '[2][A1][resumed]',
+        '[2][A1][finished]',
         '[2][A2][called]',
         '[2][A2][yielded]',
-        '[3][A2][resumed]'
+        '[3][A2][resumed]',
+        '[3][A2][finished]'
       ]);
     });
 
     test('Resource - acquire and wait', () async {
-      SimDart sim = SimDart(includeTracks: true);
-
       sim.resources.limited(name: 'r');
 
       sim.process(
@@ -63,22 +71,21 @@ void main() {
           },
           name: 'B');
 
-      SimResult result = await sim.run();
+      await sim.run();
 
-      TrackTester tt = TrackTester(result);
-      tt.test([
+      helper.test([
         '[0][A][called]',
         '[0][A][yielded]',
         '[0][B][called]',
         '[0][B][yielded]',
         '[10][A][resumed]',
-        '[10][B][resumed]'
+        '[10][A][finished]',
+        '[10][B][resumed]',
+        '[10][B][finished]'
       ]);
     });
 
     test('Resource', () async {
-      SimDart sim = SimDart(includeTracks: true);
-
       sim.resources.limited(name: 'r');
 
       sim.repeatProcess(
@@ -90,10 +97,9 @@ void main() {
           name: (start) => 'A$start',
           interval: Interval.fixed(fixedInterval: 1, untilTime: 2));
 
-      SimResult result = await sim.run();
+      await sim.run();
 
-      TrackTester tt = TrackTester(result);
-      tt.test([
+      helper.test([
         '[0][A0][called]',
         '[0][A0][yielded]',
         '[1][A1][called]',
@@ -101,18 +107,19 @@ void main() {
         '[2][A2][called]',
         '[2][A2][yielded]',
         '[10][A0][resumed]',
+        '[10][A0][finished]',
         '[10][A1][resumed]',
         '[10][A1][yielded]',
         '[20][A1][resumed]',
+        '[20][A1][finished]',
         '[20][A2][resumed]',
         '[20][A2][yielded]',
-        '[30][A2][resumed]'
+        '[30][A2][resumed]',
+        '[30][A2][finished]'
       ]);
     });
 
     test('Resource - stop', () async {
-      SimDart sim = SimDart(includeTracks: true);
-
       sim.resources.limited(name: 'r');
 
       sim.repeatProcess(
@@ -125,10 +132,14 @@ void main() {
           stopCondition: (s) => !s.resources.isAvailable('r'),
           interval: Interval.fixed(fixedInterval: 1, untilTime: 50));
 
-      SimResult result = await sim.run();
+      await sim.run();
 
-      TrackTester tt = TrackTester(result);
-      tt.test(['[0][A0][called]', '[0][A0][yielded]', '[2][A0][resumed]']);
+      helper.test([
+        '[0][A0][called]',
+        '[0][A0][yielded]',
+        '[2][A0][resumed]',
+        '[2][A0][finished]'
+      ]);
     });
   });
 }